adds conditional for recovering dir paths fixes #22
The process of creating a relative path from an absolute one seems to
strip the trailing slash, identifying a path as a directory not a file.

When this is converted into a Regex for adding to the ignore list, it
also ignores files that start with the same name as ignored directories.

This fix simply checks whether there was originally a trailing slash,
then adds it again before it is converted into a regular expression.