Anchor excludes to start of string. Fixes #44.
Before this change, excluding a directory (`bin`) would also exclude posts
that include that name (`scrubbing`), which is very unexpected.

While this may start listening to files that were previously ignored, it
more closely matches the behaviour of jekyll on it's initial build
(which _would_ include `scrubbing`).  The consequences of being overly
aggressive on listening to changes seem less important than missing
something.

This change is a bandaid. A proper fix would be to unify with Jekyll's
exclusion logic (Jekyll::EntryFilter maybe?) to ensure they will always
be the same. In the meantime, it _may_ be advisable to not include
Jekyll's `exclude` config in the watchers ignore list and take the hit
of overly aggressive reloads. I leave that decision to someone more
experienced than I.