Allow ignoring missing files
As the example of .jekyll-cache shows, it is possible to ignore missing
files/directories. However, they're currently filtered out; I don't quite
understand why. I found this limitation while playing with one of my
Jekyll projects and noticing that a directory added to "exclude" was not
ignored if it was created after `jekyll serve` is run.